# server/swagger_server/models/document_chat_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class DocumentChatResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, response: str=None):  # noqa: E501
        """DocumentChatResponse - a model defined in Swagger

        :param response: The response of this DocumentChatResponse.  # noqa: E501
        :type response: str
        """
        self.swagger_types = {
            'response': str
        }

        self.attribute_map = {
            'response': 'response'
        }
        self._response = response

    @classmethod
    def from_dict(cls, dikt) -> 'DocumentChatResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DocumentChatResponse of this DocumentChatResponse.  # noqa: E501
        :rtype: DocumentChatResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def response(self) -> str:
        """Gets the response of this DocumentChatResponse.


        :return: The response of this DocumentChatResponse.
        :rtype: str
        """
        return self._response

    @response.setter
    def response(self, response: str):
        """Sets the response of this DocumentChatResponse.


        :param response: The response of this DocumentChatResponse.
        :type response: str
        """

        self._response = response