# server/swagger_server/models/funnel_status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class FunnelStatus(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, status: str=None, count: int=None):  # noqa: E501
        """FunnelStatus - a model defined in Swagger

        :param status: The status of this FunnelStatus.  # noqa: E501
        :type status: str
        :param count: The count of this FunnelStatus.  # noqa: E501
        :type count: int
        """
        self.swagger_types = {
            'status': str,
            'count': int
        }

        self.attribute_map = {
            'status': 'status',
            'count': 'count'
        }
        self._status = status
        self._count = count

    @classmethod
    def from_dict(cls, dikt) -> 'FunnelStatus':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The FunnelStatus of this FunnelStatus.  # noqa: E501
        :rtype: FunnelStatus
        """
        return util.deserialize_model(dikt, cls)

    @property
    def status(self) -> str:
        """Gets the status of this FunnelStatus.


        :return: The status of this FunnelStatus.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this FunnelStatus.


        :param status: The status of this FunnelStatus.
        :type status: str
        """

        self._status = status

    @property
    def count(self) -> int:
        """Gets the count of this FunnelStatus.


        :return: The count of this FunnelStatus.
        :rtype: int
        """
        return self._count

    @count.setter
    def count(self, count: int):
        """Sets the count of this FunnelStatus.


        :param count: The count of this FunnelStatus.
        :type count: int
        """

        self._count = count