# server/swagger_server/models/search_query_history.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class SearchQueryHistory(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, project_id: str=None, query_id: str=None, query: str=None, timestamp: datetime=None):  # noqa: E501
        """SearchQueryHistory - a model defined in Swagger

        :param project_id: The project_id of this SearchQueryHistory.  # noqa: E501
        :type project_id: str
        :param query_id: The query_id of this SearchQueryHistory.  # noqa: E501
        :type query_id: str
        :param query: The query of this SearchQueryHistory.  # noqa: E501
        :type query: str
        :param timestamp: The timestamp of this SearchQueryHistory.  # noqa: E501
        :type timestamp: datetime
        """
        self.swagger_types = {
            'project_id': str,
            'query_id': str,
            'query': str,
            'timestamp': datetime
        }

        self.attribute_map = {
            'project_id': 'projectId',
            'query_id': 'queryId',
            'query': 'query',
            'timestamp': 'timestamp'
        }
        self._project_id = project_id
        self._query_id = query_id
        self._query = query
        self._timestamp = timestamp

    @classmethod
    def from_dict(cls, dikt) -> 'SearchQueryHistory':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SearchQueryHistory of this SearchQueryHistory.  # noqa: E501
        :rtype: SearchQueryHistory
        """
        return util.deserialize_model(dikt, cls)

    @property
    def project_id(self) -> str:
        """Gets the project_id of this SearchQueryHistory.


        :return: The project_id of this SearchQueryHistory.
        :rtype: str
        """
        return self._project_id

    @project_id.setter
    def project_id(self, project_id: str):
        """Sets the project_id of this SearchQueryHistory.


        :param project_id: The project_id of this SearchQueryHistory.
        :type project_id: str
        """

        self._project_id = project_id

    @property
    def query_id(self) -> str:
        """Gets the query_id of this SearchQueryHistory.


        :return: The query_id of this SearchQueryHistory.
        :rtype: str
        """
        return self._query_id

    @query_id.setter
    def query_id(self, query_id: str):
        """Sets the query_id of this SearchQueryHistory.


        :param query_id: The query_id of this SearchQueryHistory.
        :type query_id: str
        """

        self._query_id = query_id

    @property
    def query(self) -> str:
        """Gets the query of this SearchQueryHistory.


        :return: The query of this SearchQueryHistory.
        :rtype: str
        """
        return self._query

    @query.setter
    def query(self, query: str):
        """Sets the query of this SearchQueryHistory.


        :param query: The query of this SearchQueryHistory.
        :type query: str
        """

        self._query = query

    @property
    def timestamp(self) -> datetime:
        """Gets the timestamp of this SearchQueryHistory.


        :return: The timestamp of this SearchQueryHistory.
        :rtype: datetime
        """
        return self._timestamp

    @timestamp.setter
    def timestamp(self, timestamp: datetime):
        """Sets the timestamp of this SearchQueryHistory.


        :param timestamp: The timestamp of this SearchQueryHistory.
        :type timestamp: datetime
        """

        self._timestamp = timestamp