# server/swagger_server/models/search_result.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.search_result_pico import SearchResultPico  # noqa: F401,E501
from swagger_server import util


class SearchResult(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, title: str=None, author: str=None, year: int=None, abstract: str=None, pico: SearchResultPico=None):  # noqa: E501
        """SearchResult - a model defined in Swagger

        :param id: The id of this SearchResult.  # noqa: E501
        :type id: str
        :param title: The title of this SearchResult.  # noqa: E501
        :type title: str
        :param author: The author of this SearchResult.  # noqa: E501
        :type author: str
        :param year: The year of this SearchResult.  # noqa: E501
        :type year: int
        :param abstract: The abstract of this SearchResult.  # noqa: E501
        :type abstract: str
        :param pico: The pico of this SearchResult.  # noqa: E501
        :type pico: SearchResultPico
        """
        self.swagger_types = {
            'id': str,
            'title': str,
            'author': str,
            'year': int,
            'abstract': str,
            'pico': SearchResultPico
        }

        self.attribute_map = {
            'id': 'id',
            'title': 'title',
            'author': 'author',
            'year': 'year',
            'abstract': 'abstract',
            'pico': 'pico'
        }
        self._id = id
        self._title = title
        self._author = author
        self._year = year
        self._abstract = abstract
        self._pico = pico

    @classmethod
    def from_dict(cls, dikt) -> 'SearchResult':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SearchResult of this SearchResult.  # noqa: E501
        :rtype: SearchResult
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this SearchResult.


        :return: The id of this SearchResult.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this SearchResult.


        :param id: The id of this SearchResult.
        :type id: str
        """

        self._id = id

    @property
    def title(self) -> str:
        """Gets the title of this SearchResult.


        :return: The title of this SearchResult.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this SearchResult.


        :param title: The title of this SearchResult.
        :type title: str
        """

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this SearchResult.


        :return: The author of this SearchResult.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this SearchResult.


        :param author: The author of this SearchResult.
        :type author: str
        """

        self._author = author

    @property
    def year(self) -> int:
        """Gets the year of this SearchResult.


        :return: The year of this SearchResult.
        :rtype: int
        """
        return self._year

    @year.setter
    def year(self, year: int):
        """Sets the year of this SearchResult.


        :param year: The year of this SearchResult.
        :type year: int
        """

        self._year = year

    @property
    def abstract(self) -> str:
        """Gets the abstract of this SearchResult.


        :return: The abstract of this SearchResult.
        :rtype: str
        """
        return self._abstract

    @abstract.setter
    def abstract(self, abstract: str):
        """Sets the abstract of this SearchResult.


        :param abstract: The abstract of this SearchResult.
        :type abstract: str
        """

        self._abstract = abstract

    @property
    def pico(self) -> SearchResultPico:
        """Gets the pico of this SearchResult.


        :return: The pico of this SearchResult.
        :rtype: SearchResultPico
        """
        return self._pico

    @pico.setter
    def pico(self, pico: SearchResultPico):
        """Sets the pico of this SearchResult.


        :param pico: The pico of this SearchResult.
        :type pico: SearchResultPico
        """

        self._pico = pico